+++ gluecodium/src/test/resources/smoke/generic_types/output/android/jni/com_example_smoke_UseOptimizedListStruct__Conversion.cpp
@@ -15,10 +15,10 @@ convert_from_jni(JNIEnv* _jenv, const JniReference<jobject>& _jinput, ::smoke::U
 {
     auto j_structs = get_object_field_value(_jenv, _jinput, "structs", "Ljava/util/List;");
     auto structs_handle = get_field_value(_jenv, j_structs, "nativeHandle", (int64_t*)nullptr);
-    auto& n_structs = *reinterpret_cast<std::shared_ptr<::std::vector< ::std::shared_ptr< ::smoke::VeryBigStruct > >>*>(structs_handle);
+    auto& n_structs = **reinterpret_cast<std::shared_ptr<::std::vector< ::std::shared_ptr< ::smoke::VeryBigStruct > >>*>(structs_handle);
     auto j_classes = get_object_field_value(_jenv, _jinput, "classes", "Ljava/util/List;");
     auto classes_handle = get_field_value(_jenv, j_classes, "nativeHandle", (int64_t*)nullptr);
-    auto& n_classes = *reinterpret_cast<std::shared_ptr<::std::vector< ::std::shared_ptr< ::smoke::UnreasonablyLazyClass > >>*>(classes_handle);
+    auto& n_classes = **reinterpret_cast<std::shared_ptr<::std::vector< ::std::shared_ptr< ::smoke::UnreasonablyLazyClass > >>*>(classes_handle);
     return ::smoke::UseOptimizedListStruct(std::move(n_structs), std::move(n_classes));
 }
 ::gluecodium::optional<::smoke::UseOptimizedListStruct>
@@ -34,9 +34,9 @@ convert_to_jni(JNIEnv* _jenv, const ::smoke::UseOptimizedListStruct& _ninput)
 {
     auto& javaClass = CachedJavaClass<::smoke::UseOptimizedListStruct>::java_class;
     auto _jresult = ::gluecodium::jni::alloc_object(_jenv, javaClass);
-    auto jstructs = convert_to_jni_optimized(_jenv, result, "com/example/smoke/UseOptimizedListStruct$VeryBigStructLazyNativeList");
+    auto jstructs = convert_to_jni_optimized(_jenv, _ninput.structs, "com/example/smoke/UseOptimizedListStruct$VeryBigStructLazyNativeList");
     ::gluecodium::jni::set_object_field_value(_jenv, _jresult, "structs", "Ljava/util/List;", jstructs);
-    auto jclasses = convert_to_jni_optimized(_jenv, result, "com/example/smoke/UseOptimizedListStruct$UnreasonablyLazyClassLazyNativeList");
+    auto jclasses = convert_to_jni_optimized(_jenv, _ninput.classes, "com/example/smoke/UseOptimizedListStruct$UnreasonablyLazyClassLazyNativeList");
     ::gluecodium::jni::set_object_field_value(_jenv, _jresult, "classes", "Ljava/util/List;", jclasses);
     return _jresult;
 }